pbs-final-v3/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1
prometheus-client==0.20.0
psutil==6.0.0
PyYAML==6.0.2

streamlit==1.38.0
pandas==2.2.2
plotly==5.23.0
